--- plugin-taskbar/backends/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_subdirectory(wayland)
-add_subdirectory(xcb)
