+++ plugin-taskbar/lxqttaskgroup.cpp
@@ -41,7 +41,7 @@
 #include <QMenu>
 #include <XdgIcon>
 
-#include "ilxqttaskbarabstractbackend.h"
+#include "../panel/backends/ilxqttaskbarabstractbackend.h"
 
 /************************************************
